pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sci-llm"
description = ""
version = "0.0.0"
requires-python = "~=3.13.0"
authors = [
    {name="Albert Gong", email="agong@cs.cornell.edu"},
]
readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "huggingface_hub",
    "datasets",
    "pdfplumber",
    "langchain",
    "pandas",
    "pydantic",
    "pymatgen",
    "python-dotenv",
    "pyyaml",
    "tqdm",
    "google-genai",
    "openai",
    "pymupdf",
    "scikit-learn>=1.8.0",
    "python-slugify>=8.0.4",
]

[dependency-groups]
dev = [
    "harbor>=0.1.23",
    "pre-commit",
    "ruff",
]
validator = [
    "streamlit",
    "pymupdf",
]

[project.urls]

[project.scripts]
pbench-extract = "pbench.mass_extract_properties_from_llm:main"
pbench-filter = "pbench.filter_candidates:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
pbench = ["*.yaml"]
